+++ crates/air/src/layout/starknet_with_keccak/mod.rs
@@ -121,15 +121,15 @@ pub const BUILTINS: [Felt; 8] = [
 pub struct Layout {}
 
 impl StaticLayoutTrait for Layout {
-    const NUM_COLUMNS_FIRST: usize = 12;
-    const NUM_COLUMNS_SECOND: usize = 3;
+    const NUM_COLUMNS_FIRST: u32 = 12;
+    const NUM_COLUMNS_SECOND: u32 = 3;
 }
 
 impl GenericLayoutTrait for Layout {
-    fn get_num_columns_first(_public_input: &PublicInput) -> Option<usize> {
+    fn get_num_columns_first(_public_input: &PublicInput) -> Option<u32> {
         Some(Self::NUM_COLUMNS_FIRST)
     }
-    fn get_num_columns_second(_public_input: &PublicInput) -> Option<usize> {
+    fn get_num_columns_second(_public_input: &PublicInput) -> Option<u32> {
         Some(Self::NUM_COLUMNS_SECOND)
     }
 }
